+++ QuizApp/QuizApp/UI/Common/QuizList/AppData.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+class AppData: ObservableObject {
+
+    @Published var loginStatus: LoginStatus = .unknown
+    @Published var selectedTab: AppRoute = .quizzes
+
+}
+
+enum LoginStatus {
+
+    case unknown
+    case notLoggedIn
+    case loggedIn
+
+}
