+++ QuizApp/QuizApp/UI/Screens/ContentView.swift
@@ -40,18 +40,3 @@ struct ContentView: View {
     }
 
 }
-
-class AppData: ObservableObject {
-
-    @Published var loginStatus: LoginStatus = .unknown
-    @Published var selectedTab: AppRoute = .quizzes
-
-}
-
-enum LoginStatus {
-
-    case unknown
-    case notLoggedIn
-    case loggedIn
-
-}
